Fix a problem where an error in onItem could result in a bad connection being reused
If an error occurs inside the fn() that onItem calls it would have not been spotted by
condRelease() and that could result in a bad connection being placed back in the
pool of free connections. Under load the package was experiencing this problem manifested
in the following fashion:

  1. A connection gets an i/o timeout during a call to onItem()

  2. The error returned by the function called by onItem() is placed in a new
     err value (because of the :=) but the condRelease() has been set up with
     &err from a previous declaration of err. Therefore err is nil.

  3. condRelease() places the connection back in the free list.

  4. The connection is reused at some point and there's data waiting to be read
     from the previous memcache command that got the i/o timeout. This results
     in (typically) an error from the Sscanf line at line 405 because that line
     ends up trying to parse part of the response that was not previously read
     because of the i/o timeout.

This fix prevents this from happening. If an error occurs the connection will not be
reused.