Attempt to drain error'd connections.
If a connection times out, instead of closing it run a goroutine to drain anything thats left on the connection and then close it.

We occasionally see lots (10s of thousands) of sockets on the memcached server in CLOSE_WAIT, all with data in their send buffers.  This looks like a bug in memcache, and this is a work around on the client side.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>